+++ test/session.js
@@ -2,22 +2,23 @@
 process.env.NO_DEPRECATION = 'express-session';
 
 var after = require('after')
+var assert = require('assert')
 var express = require('express')
-  , assert = require('assert')
   , request = require('supertest')
-  , should = require('should')
   , cookieParser = require('cookie-parser')
   , session = require('../')
   , Cookie = require('../session/cookie')
-var http = require('http');
+var fs = require('fs')
+var http = require('http')
+var https = require('https')
 
 var min = 60 * 1000;
 
 describe('session()', function(){
   it('should export constructors', function(){
-    session.Session.should.be.a.Function;
-    session.Store.should.be.a.Function;
-    session.MemoryStore.should.be.a.Function;
+    assert.equal(typeof session.Session, 'function')
+    assert.equal(typeof session.Store, 'function')
+    assert.equal(typeof session.MemoryStore, 'function')
   })
 
   it('should do nothing if req.session exists', function(done){
@@ -28,11 +29,8 @@ describe('session()', function(){
 
       request(app)
       .get('/')
-      .expect(200, function(err, res){
-        if (err) return done(err);
-        should(cookie(res)).be.empty;
-        done();
-      });
+      .expect(shouldNotHaveHeader('Set-Cookie'))
+      .expect(200, done)
   })
 
   it('should error without secret', function(done){
@@ -48,9 +46,9 @@ describe('session()', function(){
       .use(end);
     app.set('env', 'test');
 
-      request(app)
-      .get('/')
-      .expect(200, '', done);
+    request(app)
+    .get('/')
+    .expect(200, '', done)
   })
 
   it('should create a new session', function (done) {
@@ -62,12 +60,12 @@ describe('session()', function(){
 
     request(server)
     .get('/')
+    .expect(shouldSetCookie('connect.sid'))
     .expect(200, 'session active', function (err, res) {
       if (err) return done(err)
-      should(sid(res)).not.be.empty
       store.length(function (err, len) {
         if (err) return done(err)
-        len.should.equal(1)
+        assert.equal(len, 1)
         done()
       })
     })
@@ -82,9 +80,9 @@ describe('session()', function(){
 
     request(server)
     .get('/')
+    .expect(shouldSetCookie('connect.sid'))
     .expect(200, 'session 1', function (err, res) {
       if (err) return done(err)
-      should(sid(res)).not.be.empty
       request(server)
       .get('/')
       .set('Cookie', cookie(res))
@@ -140,9 +138,9 @@ describe('session()', function(){
 
     request(server)
     .get('/')
+    .expect(shouldSetCookie('connect.sid'))
     .expect(200, 'hello, world', function (err, res) {
       if (err) return done(err)
-      should(sid(res)).not.be.empty
       request(server)
       .get('/')
       .set('Cookie', cookie(res))
@@ -166,9 +164,9 @@ describe('session()', function(){
 
     request(server)
     .get('/')
+    .expect(shouldSetCookie('connect.sid'))
     .expect(200, 'session 1', function (err, res) {
       if (err) return done(err)
-      should(sid(res)).not.be.empty
       request(server)
       .get('/')
       .set('Cookie', cookie(res))
@@ -190,7 +188,7 @@ describe('session()', function(){
       if (err) return done(err)
       store.all(function (err, sess) {
         if (err) return done(err)
-        Object.keys(sess).should.have.length(2)
+        assert.equal(Object.keys(sess).length, 2)
         done()
       })
     }
@@ -271,6 +269,28 @@ describe('session()', function(){
     .expect(200, 'cookie,test1,test2', done)
   })
 
+  it('should not save with bogus req.sessionID', function (done) {
+    var store = new session.MemoryStore()
+    var server = createServer({ store: store }, function (req, res) {
+      req.sessionID = function () {}
+      req.session.test1 = 1
+      req.session.test2 = 'b'
+      res.end()
+    })
+
+    request(server)
+    .get('/')
+    .expect(shouldNotHaveHeader('Set-Cookie'))
+    .expect(200, function (err) {
+      if (err) return done(err)
+      store.length(function (err, length) {
+        if (err) return done(err)
+        assert.equal(length, 0)
+        done()
+      })
+    })
+  })
+
   describe('when response ended', function () {
     it('should have saved session', function (done) {
       var saved = false
@@ -294,7 +314,7 @@ describe('session()', function(){
       .get('/')
       .expect(200, 'session saved', function (err) {
         if (err) return done(err)
-        saved.should.be.true
+        assert.ok(saved)
         done()
       })
     })
@@ -322,7 +342,7 @@ describe('session()', function(){
       .get('/')
       .expect(200, '', function (err) {
         if (err) return done(err)
-        saved.should.be.true
+        assert.ok(saved)
         done()
       })
     })
@@ -351,7 +371,7 @@ describe('session()', function(){
       .get('/')
       .expect(200, 'hello, world', function (err) {
         if (err) return done(err)
-        saved.should.be.true
+        assert.ok(saved)
         done()
       })
     })
@@ -379,7 +399,7 @@ describe('session()', function(){
       .get('/')
       .expect(200, 'session saved', function (err) {
         if (err) return done(err)
-        saved.should.be.true
+        assert.ok(saved)
         done()
       })
     })
@@ -396,9 +416,9 @@ describe('session()', function(){
 
       request(server)
       .get('/')
+      .expect(shouldSetCookie('connect.sid'))
       .expect(200, 'session 1', function (err, res) {
         if (err) return done(err)
-        should(sid(res)).not.be.empty
         store.clear(function (err) {
           if (err) return done(err)
           request(server)
@@ -419,19 +439,20 @@ describe('session()', function(){
 
       request(server)
       .get('/')
+      .expect(shouldSetCookie('connect.sid'))
       .expect(200, 'session 1', function (err, res) {
         if (err) return done(err)
         var val = sid(res)
-        should(val).not.be.empty
+        assert.ok(val)
         store.clear(function (err) {
           if (err) return done(err)
           request(server)
           .get('/')
           .set('Cookie', cookie(res))
+          .expect(shouldSetCookie('connect.sid'))
           .expect(200, 'session 2', function (err, res) {
             if (err) return done(err)
-            should(sid(res)).not.be.empty
-            should(sid(res)).not.equal(val)
+            assert.notEqual(sid(res), val)
             done()
           })
         })
@@ -450,13 +471,15 @@ describe('session()', function(){
 
       request(server)
       .get('/')
+      .expect(shouldSetCookie('sessid'))
       .expect(200, 'session created', function (err, res) {
         if (err) return done(err)
         var val = sid(res)
-        should(val).not.be.empty
+        assert.ok(val)
         request(server)
         .get('/')
         .set('Cookie', 'sessid=' + val)
+        .expect(shouldSetCookie('sessid'))
         .expect(200, 'session created', done)
       })
     })
@@ -471,14 +494,16 @@ describe('session()', function(){
 
       request(server)
       .get('/')
+      .expect(shouldSetCookie('sessid'))
       .expect(200, 'session created', function (err, res) {
         if (err) return done(err)
         var val = cookie(res).replace(/...\./, '.')
 
-        should(val).not.be.empty
+        assert.ok(val)
         request(server)
         .get('/')
         .set('Cookie', val)
+        .expect(shouldSetCookie('sessid'))
         .expect(200, 'session created', done)
       })
     })
@@ -495,15 +520,16 @@ describe('session()', function(){
 
       request(server)
       .get('/')
+      .expect(shouldSetCookie('connect.sid'))
       .expect(200, 'session 1', function (err, res) {
         if (err) return done(err)
-        should(sid(res)).not.be.empty
         setTimeout(function () {
           request(server)
           .get('/')
           .set('Cookie', cookie(res))
+          .expect(shouldSetCookie('connect.sid'))
           .expect(200, 'session 2', done)
-        }, 10)
+        }, 20)
       })
     })
 
@@ -517,18 +543,19 @@ describe('session()', function(){
 
       request(server)
       .get('/')
+      .expect(shouldSetCookie('connect.sid'))
       .expect(200, 'session 1', function (err, res) {
         if (err) return done(err)
         var val = sid(res)
-        should(val).not.be.empty
+        assert.ok(val)
         setTimeout(function () {
           request(server)
           .get('/')
           .set('Cookie', cookie(res))
+          .expect(shouldSetCookie('connect.sid'))
           .expect(200, 'session 2', function (err, res) {
             if (err) return done(err)
-            should(sid(res)).not.be.empty
-            should(sid(res)).not.equal(val)
+            assert.notEqual(sid(res), val)
             done()
           })
         }, 15)
@@ -545,14 +572,13 @@ describe('session()', function(){
 
       request(server)
       .get('/')
+      .expect(shouldSetCookie('connect.sid'))
       .expect(200, 'session 1', function (err, res) {
         if (err) return done(err)
-        var val = sid(res)
-        should(val).not.be.empty
         setTimeout(function () {
           store.all(function (err, sess) {
             if (err) return done(err)
-            Object.keys(sess).should.have.length(0)
+            assert.equal(Object.keys(sess).length, 0)
             done()
           })
         }, 10)
@@ -571,32 +597,23 @@ describe('session()', function(){
         request(server)
         .get('/')
         .set('X-Forwarded-Proto', 'https')
-        .expect(200, function(err, res){
-          if (err) return done(err);
-          should(cookie(res)).not.be.empty;
-          done();
-        });
+        .expect(shouldSetCookie('connect.sid'))
+        .expect(200, done)
       })
 
       it('should trust X-Forwarded-Proto when comma-separated list', function(done){
         request(server)
         .get('/')
         .set('X-Forwarded-Proto', 'https,http')
-        .expect(200, function(err, res){
-          if (err) return done(err);
-          should(cookie(res)).not.be.empty;
-          done();
-        });
+        .expect(shouldSetCookie('connect.sid'))
+        .expect(200, done)
       })
 
       it('should work when no header', function(done){
         request(server)
         .get('/')
-        .expect(200, function(err, res){
-          if (err) return done(err);
-          should(cookie(res)).be.empty;
-          done();
-        });
+        .expect(shouldNotHaveHeader('Set-Cookie'))
+        .expect(200, done)
       })
     })
 
@@ -610,11 +627,8 @@ describe('session()', function(){
         request(server)
         .get('/')
         .set('X-Forwarded-Proto', 'https')
-        .expect(200, function(err, res){
-          if (err) return done(err);
-          should(cookie(res)).be.empty;
-          done();
-        });
+        .expect(shouldNotHaveHeader('Set-Cookie'))
+        .expect(200, done)
       })
 
       it('should ignore req.secure from express', function(done){
@@ -626,11 +640,8 @@ describe('session()', function(){
         request(app)
         .get('/')
         .set('X-Forwarded-Proto', 'https')
-        .expect(200, 'true', function(err, res){
-          if (err) return done(err);
-          should(cookie(res)).be.empty;
-          done();
-        });
+        .expect(shouldNotHaveHeader('Set-Cookie'))
+        .expect(200, 'true', done)
       })
     })
 
@@ -644,11 +655,8 @@ describe('session()', function(){
         request(server)
         .get('/')
         .set('X-Forwarded-Proto', 'https')
-        .expect(200, function(err, res){
-          if (err) return done(err);
-          should(cookie(res)).be.empty;
-          done();
-        });
+        .expect(shouldNotHaveHeader('Set-Cookie'))
+        .expect(200, done)
       })
 
       it('should use req.secure from express', function(done){
@@ -660,28 +668,22 @@ describe('session()', function(){
         request(app)
         .get('/')
         .set('X-Forwarded-Proto', 'https')
-        .expect(200, 'true', function(err, res){
-          if (err) return done(err);
-          should(cookie(res)).not.be.empty;
-          done();
-        });
+        .expect(shouldSetCookie('connect.sid'))
+        .expect(200, 'true', done)
       })
     })
   })
 
   describe('genid option', function(){
     it('should reject non-function values', function(){
-      session.bind(null, { genid: 'bogus!' }).should.throw(/genid.*must/);
+      assert.throws(session.bind(null, { genid: 'bogus!' }), /genid.*must/)
     });
 
     it('should provide default generator', function(done){
       request(createServer())
       .get('/')
-      .expect(200, function (err, res) {
-        if (err) return done(err)
-        should(sid(res)).not.be.empty
-        done()
-      })
+      .expect(shouldSetCookie('connect.sid'))
+      .expect(200, done)
     });
 
     it('should allow custom function', function(done){
@@ -689,11 +691,8 @@ describe('session()', function(){
 
       request(createServer({ genid: genid }))
       .get('/')
-      .expect(200, function (err, res) {
-        if (err) return done(err)
-        should(sid(res)).equal('apple')
-        done()
-      })
+      .expect(shouldSetCookieToValue('connect.sid', 's%3Aapple.D8Y%2BpkTAmeR0PobOhY4G97PRW%2Bj7bUnP%2F5m6%2FOn1MCU'))
+      .expect(200, done)
     });
 
     it('should encode unsafe chars', function(done){
@@ -701,11 +700,8 @@ describe('session()', function(){
 
       request(createServer({ genid: genid }))
       .get('/')
-      .expect(200, function (err, res) {
-        if (err) return done(err)
-        should(sid(res)).equal('%25')
-        done()
-      })
+      .expect(shouldSetCookieToValue('connect.sid', 's%3A%25.kzQ6x52kKVdF35Qh62AWk4ZekS28K5XYCXKa%2FOTZ01g'))
+      .expect(200, done)
     });
 
     it('should provide req argument', function(done){
@@ -713,11 +709,8 @@ describe('session()', function(){
 
       request(createServer({ genid: genid }))
       .get('/foo')
-      .expect(200, function (err, res) {
-        if (err) return done(err)
-        should(sid(res)).equal('%2Ffoo')
-        done()
-      })
+      .expect(shouldSetCookieToValue('connect.sid', 's%3A%2Ffoo.paEKBtAHbV5s1IB8B2zPnzAgYmmnRPIqObW4VRYj%2FMQ'))
+      .expect(200, done)
     });
   });
 
@@ -725,21 +718,15 @@ describe('session()', function(){
     it('should default to "connect.sid"', function(done){
       request(createServer())
       .get('/')
-      .end(function(err, res){
-        res.headers['set-cookie'].should.have.length(1);
-        res.headers['set-cookie'][0].should.match(/^connect\.sid/);
-        done();
-      });
+      .expect(shouldSetCookie('connect.sid'))
+      .expect(200, done)
     })
 
     it('should allow overriding', function(done){
-      request(createServer({ key: 'sid' }))
+      request(createServer({ key: 'session_id' }))
       .get('/')
-      .end(function(err, res){
-        res.headers['set-cookie'].should.have.length(1);
-        res.headers['set-cookie'][0].should.match(/^sid/);
-        done();
-      });
+      .expect(shouldSetCookie('session_id'))
+      .expect(200, done)
     })
   })
 
@@ -755,17 +742,14 @@ describe('session()', function(){
 
       request(app)
       .get('/')
+      .expect(shouldSetCookie('connect.sid'))
       .expect(200, function(err, res){
         if (err) return done(err);
-        should(cookie(res)).not.be.empty;
         request(app)
         .get('/')
         .set('Cookie', cookie(res))
-        .expect(200, function(err, res){
-          if (err) return done(err);
-          should(cookie(res)).be.empty;
-          done();
-        });
+        .expect(shouldNotHaveHeader('Set-Cookie'))
+        .expect(200, done)
       });
     });
 
@@ -780,17 +764,14 @@ describe('session()', function(){
 
       request(app)
       .get('/')
+      .expect(shouldSetCookie('connect.sid'))
       .expect(200, function(err, res){
         if (err) return done(err);
-        should(cookie(res)).not.be.empty;
         request(app)
         .get('/')
         .set('Cookie', cookie(res))
-        .expect(200, function(err, res){
-          if (err) return done(err);
-          should(cookie(res)).not.be.empty;
-          done();
-        });
+        .expect(shouldSetCookie('connect.sid'))
+        .expect(200, done)
       });
     });
   });
@@ -908,6 +889,35 @@ describe('session()', function(){
         .expect(200, done);
       });
     });
+
+    it('should pass session touch error', function (done) {
+      var cb = after(2, done)
+      var store = new session.MemoryStore()
+      var server = createServer({ store: store, resave: false }, function (req, res) {
+        req.session.hit = true
+        res.end('session saved')
+      })
+
+      store.touch = function touch(sid, sess, callback) {
+        callback(new Error('boom!'))
+      }
+
+      server.on('error', function onerror(err) {
+        assert.ok(err)
+        assert.equal(err.message, 'boom!')
+        cb()
+      })
+
+      request(server)
+      .get('/')
+      .expect(200, 'session saved', function (err, res) {
+        if (err) return cb(err)
+        request(server)
+        .get('/')
+        .set('Cookie', cookie(res))
+        .end(cb)
+      })
+    })
   });
 
   describe('saveUninitialized option', function(){
@@ -928,7 +938,7 @@ describe('session()', function(){
       request(app)
       .get('/')
       .expect('x-count', '1')
-      .expect('set-cookie', /connect\.sid=/)
+      .expect(shouldSetCookie('connect.sid'))
       .expect(200, done);
     });
 
@@ -949,7 +959,7 @@ describe('session()', function(){
       request(app)
       .get('/')
       .expect('x-count', '1')
-      .expect('set-cookie', /connect\.sid=/)
+      .expect(shouldSetCookie('connect.sid'))
       .expect(200, done);
     });
 
@@ -970,11 +980,8 @@ describe('session()', function(){
       request(app)
       .get('/')
       .expect('x-count', '0')
-      .expect(200, function(err, res){
-        if (err) return done(err);
-        should(cookie(res)).be.empty;
-        done();
-      });
+      .expect(shouldNotHaveHeader('Set-Cookie'))
+      .expect(200, done)
     });
 
     it('should still save modified session', function(done){
@@ -996,7 +1003,7 @@ describe('session()', function(){
       request(app)
       .get('/')
       .expect('x-count', '1')
-      .expect('set-cookie', /connect\.sid=/)
+      .expect(shouldSetCookie('connect.sid'))
       .expect(200, done);
     });
 
@@ -1012,7 +1019,8 @@ describe('session()', function(){
       }
 
       server.on('error', function onerror(err) {
-        err.message.should.equal('boom!')
+        assert.ok(err)
+        assert.equal(err.message, 'boom!')
         cb()
       })
 
@@ -1020,11 +1028,27 @@ describe('session()', function(){
       .get('/')
       .expect(200, 'session saved', cb)
     })
+
+    it('should prevent uninitialized session from being touched', function (done) {
+      var cb = after(1, done)
+      var store = new session.MemoryStore()
+      var server = createServer({ saveUninitialized: false, store: store, cookie: { maxAge: min } }, function (req, res) {
+        res.end()
+      })
+
+      store.touch = function () {
+        cb(new Error('should not be called'))
+      }
+
+      request(server)
+      .get('/')
+      .expect(200, cb)
+    })
   });
 
   describe('unset option', function () {
     it('should reject unknown values', function(){
-      session.bind(null, { unset: 'bogus!' }).should.throw(/unset.*must/);
+      assert.throws(session.bind(null, { unset: 'bogus!' }), /unset.*must/)
     });
 
     it('should default to keep', function(done){
@@ -1044,7 +1068,7 @@ describe('session()', function(){
         if (err) return done(err);
         store.length(function(err, len){
           if (err) return done(err);
-          len.should.equal(1);
+          assert.equal(len, 1)
           request(app)
           .get('/')
           .set('Cookie', cookie(res))
@@ -1052,7 +1076,7 @@ describe('session()', function(){
             if (err) return done(err);
             store.length(function(err, len){
               if (err) return done(err);
-              len.should.equal(1);
+              assert.equal(len, 1)
               done();
             });
           });
@@ -1077,7 +1101,7 @@ describe('session()', function(){
         if (err) return done(err);
         store.length(function(err, len){
           if (err) return done(err);
-          len.should.equal(1);
+          assert.equal(len, 1)
           request(app)
           .get('/')
           .set('Cookie', cookie(res))
@@ -1085,7 +1109,7 @@ describe('session()', function(){
             if (err) return done(err);
             store.length(function(err, len){
               if (err) return done(err);
-              len.should.equal(0);
+              assert.equal(len, 0)
               done();
             });
           });
@@ -1104,12 +1128,12 @@ describe('session()', function(){
 
       request(app)
       .get('/')
+      .expect(shouldNotHaveHeader('Set-Cookie'))
       .expect(200, function(err, res){
         if (err) return done(err);
         store.length(function(err, len){
           if (err) return done(err);
-          len.should.equal(0);
-          should(cookie(res)).be.empty;
+          assert.equal(len, 0)
           done();
         });
       });
@@ -1128,7 +1152,8 @@ describe('session()', function(){
       }
 
       server.on('error', function onerror(err) {
-        err.message.should.equal('boom!')
+        assert.ok(err)
+        assert.equal(err.message, 'boom!')
         cb()
       })
 
@@ -1187,7 +1212,7 @@ describe('session()', function(){
         if (err) return done(err)
         store.load(sid(res), function (err, sess) {
           if (err) return done(err)
-          should(sess).not.be.empty
+          assert.ok(sess)
           request(server)
           .get('/')
           .set('Cookie', cookie(res))
@@ -1211,33 +1236,29 @@ describe('session()', function(){
 
       request(app)
       .get('/')
-      .end(function(err, res){
-        res.text.should.equal('1');
-
+      .expect(200, '1', function (err, res) {
+        if (err) return done(err)
         request(app)
         .get('/')
         .set('Cookie', cookie(res))
-        .end(function(err, res){
+        .expect(200, '2', function (err, res) {
+          if (err) return done(err)
           var val = cookie(res);
-          res.text.should.equal('2');
           modify = false;
 
           request(app)
           .get('/')
           .set('Cookie', val)
-          .end(function(err, res){
-            should(sid(res)).be.empty;
-            res.text.should.equal('2');
+          .expect(shouldNotHaveHeader('Set-Cookie'))
+          .expect(200, '2', function (err, res) {
+            if (err) return done(err)
             modify = true;
 
             request(app)
             .get('/')
             .set('Cookie', val)
-            .end(function(err, res){
-              sid(res).should.not.be.empty;
-              res.text.should.equal('3');
-              done();
-            });
+            .expect(shouldSetCookie('connect.sid'))
+            .expect(200, '3', done)
           });
         });
       });
@@ -1257,10 +1278,8 @@ describe('session()', function(){
 
         request(app)
         .get('/')
-        .end(function(err, res){
-          res.headers.should.not.have.property('set-cookie');
-          done();
-        });
+        .expect(shouldNotHaveHeader('Set-Cookie'))
+        .expect(200, done)
       })
     })
 
@@ -1272,23 +1291,24 @@ describe('session()', function(){
             var id = req.session.id;
             req.session.regenerate(function(err){
               if (err) throw err;
-              id.should.not.equal(req.session.id);
+              assert.notEqual(id, req.session.id)
               res.end();
             });
           });
 
         request(app)
         .get('/')
-        .end(function(err, res){
+        .expect(shouldSetCookie('connect.sid'))
+        .expect(200, function (err, res) {
           if (err) return done(err)
           var id = sid(res)
           request(app)
           .get('/')
           .set('Cookie', cookie(res))
-          .end(function(err, res){
+          .expect(shouldSetCookie('connect.sid'))
+          .expect(200, function (err, res) {
             if (err) return done(err)
-            should(sid(res)).not.be.empty
-            should(sid(res)).should.not.equal(id)
+            assert.notEqual(sid(res), id)
             done();
           });
         });
@@ -1406,15 +1426,50 @@ describe('session()', function(){
         .get('/')
         .expect(200, 'saved', function (err, res) {
           if (err) return done(err)
-          count.should.equal(1)
+          assert.equal(count, 1)
           count = 0
           request(server)
           .get('/')
           .set('Cookie', cookie(res))
           .expect(200, 'saved', function (err) {
             if (err) return done(err)
-            count.should.equal(1)
+            assert.equal(count, 1)
             done()
+          })
+        })
+      })
+    })
+
+    describe('.touch()', function () {
+      it('should reset session expiration', function (done) {
+        var store = new session.MemoryStore()
+        var server = createServer({ resave: false, store: store, cookie: { maxAge: min } }, function (req, res) {
+          req.session.hit = true
+          req.session.touch()
+          res.end()
+        })
+
+        request(server)
+        .get('/')
+        .expect(200, function (err, res) {
+          if (err) return done(err)
+          var id = sid(res)
+          store.get(id, function (err, sess) {
+            if (err) return done(err)
+            var exp = new Date(sess.cookie.expires)
+            setTimeout(function () {
+              request(server)
+              .get('/')
+              .set('Cookie', cookie(res))
+              .expect(200, function (err, res) {
+                if (err) return done(err);
+                store.get(id, function (err, sess) {
+                  if (err) return done(err)
+                  assert.notEqual(new Date(sess.cookie.expires).getTime(), exp.getTime())
+                  done()
+                })
+              })
+            }, 100)
           })
         })
       })
@@ -1437,8 +1492,8 @@ describe('session()', function(){
           .expect(200, function(err, res){
             if (err) return done(err);
             var val = cookie(res);
-            should(val).not.containEql('HttpOnly');
-            should(val).containEql('Secure');
+            assert.equal(val.indexOf('HttpOnly'), -1, 'should not be HttpOnly cookie')
+            assert.notEqual(val.indexOf('Secure'), -1, 'should be Secure cookie')
             done();
           });
         })
@@ -1455,7 +1510,7 @@ describe('session()', function(){
           .expect(200, function(err, res){
             if (err) return done(err);
             var val = cookie(res);
-            should(val).not.containEql('Expires');
+            assert.equal(val.indexOf('Expires'), -1, 'should be not have cookie Expires')
             done();
           });
         })
@@ -1469,16 +1524,14 @@ describe('session()', function(){
 
           request(app)
           .get('/admin/foo')
-          .end(function(err, res){
-            res.headers.should.have.property('set-cookie');
-
+          .expect(shouldSetCookie('connect.sid'))
+          .expect(200, function (err, res) {
+            if (err) return done(err)
             request(app)
             .get('/admin')
             .set('Cookie', cookie(res))
-            .end(function(err, res){
-              res.headers.should.not.have.property('set-cookie');
-              done();
-            })
+            .expect(shouldNotHaveHeader('Set-Cookie'))
+            .expect(200, done)
           });
         })
 
@@ -1495,21 +1548,15 @@ describe('session()', function(){
           .expect(200, function(err, res){
             if (err) return done(err);
             var val = cookie(res);
-            should(val).not.containEql('HttpOnly');
-            should(val).not.containEql('Secure');
-            should(val).containEql('Path=/admin');
-            should(val).containEql('Expires');
+            assert.equal(val.indexOf('HttpOnly'), -1, 'should not be HttpOnly cookie')
+            assert.equal(val.indexOf('Secure'), -1, 'should not be Secure cookie')
+            assert.notEqual(val.indexOf('Path=/admin'), -1, 'should have cookie path /admin')
+            assert.notEqual(val.indexOf('Expires'), -1, 'should have cookie Expires')
             done();
           });
         })
 
         it('should preserve cookies set before writeHead is called', function(done){
-          function getPreviousCookie(res) {
-            var val = res.headers['set-cookie'];
-            if (!val) return '';
-            return /previous=([^;]+);/.exec(val[0])[1];
-          }
-
           var app = express()
             .use(session({ secret: 'keyboard cat' }))
             .use(function(req, res, next){
@@ -1520,28 +1567,48 @@ describe('session()', function(){
 
           request(app)
           .get('/')
-          .end(function(err, res){
-            getPreviousCookie(res).should.equal('cookieValue');
-            done();
-          });
+          .expect(shouldSetCookieToValue('previous', 'cookieValue'))
+          .expect(200, done)
         })
       })
 
       describe('.secure', function(){
-        it('should not set-cookie when insecure', function(done){
-          var app = express()
-            .use(session({ secret: 'keyboard cat' }))
-            .use(function(req, res, next){
-              req.session.cookie.secure = true;
-              res.end();
-            });
+        var app
 
-          request(app)
+        before(function () {
+          app = createRequestListener({ secret: 'keyboard cat', cookie: { secure: true } })
+        })
+
+        it('should set cookie when secure', function (done) {
+          var cert = fs.readFileSync(__dirname + '/fixtures/server.crt', 'ascii')
+          var server = https.createServer({
+            key: fs.readFileSync(__dirname + '/fixtures/server.key', 'ascii'),
+            cert: cert
+          })
+
+          server.on('request', app)
+
+          var agent = new https.Agent({ca: cert})
+          var createConnection = agent.createConnection
+
+          agent.createConnection = function (options) {
+            options.servername = 'express-session.local'
+            return createConnection.call(this, options)
+          }
+
+          var req = request(server).get('/')
+          req.agent(agent)
+          req.expect(shouldSetCookie('connect.sid'))
+          req.expect(200, done)
+        })
+
+        it('should not set-cookie when insecure', function(done){
+          var server = http.createServer(app)
+
+          request(server)
           .get('/')
-          .end(function(err, res){
-            res.headers.should.not.have.property('set-cookie');
-            done();
-          });
+          .expect(shouldNotHaveHeader('Set-Cookie'))
+          .expect(200, done)
         })
       })
 
@@ -1559,11 +1626,8 @@ describe('session()', function(){
 
           request(app)
           .get('/')
-          .end(function(err, res){
-            res.status.should.equal(200);
-            res.headers.should.not.have.property('set-cookie');
-            done();
-          });
+          .expect(shouldNotHaveHeader('Set-Cookie'))
+          .expect(200, done)
         })
 
         it('should not set-cookie even for FQDN', function(done){
@@ -1581,11 +1645,8 @@ describe('session()', function(){
           request(app)
           .get('/')
           .set('host', 'http://foo/bar')
-          .end(function(err, res){
-            res.status.should.equal(200);
-            res.headers.should.not.have.property('set-cookie');
-            done();
-          });
+          .expect(shouldNotHaveHeader('Set-Cookie'))
+          .expect(200, done)
         })
       })
 
@@ -1600,10 +1661,8 @@ describe('session()', function(){
 
           request(app)
           .get('/foo/bar/baz')
-          .end(function(err, res){
-            res.headers.should.have.property('set-cookie');
-            done();
-          });
+          .expect(shouldSetCookie('connect.sid'))
+          .expect(200, done)
         })
 
         it('should set-cookie even for FQDN', function(done){
@@ -1617,11 +1676,8 @@ describe('session()', function(){
           request(app)
           .get('/foo/bar/baz')
           .set('host', 'http://example.com')
-          .end(function(err, res){
-            res.status.should.equal(200);
-            res.headers.should.have.property('set-cookie');
-            done();
-          });
+          .expect(shouldSetCookie('connect.sid'))
+          .expect(200, done)
         })
       })
 
@@ -1640,19 +1696,14 @@ describe('session()', function(){
         it('should set relative in milliseconds', function(done){
           request(app)
           .get('/')
-          .end(function(err, res){
+          .expect(200, '1', function (err, res) {
             var a = new Date(expires(res))
-              , b = new Date;
+            var b = new Date
+            var delta = a.valueOf() - b.valueOf()
 
             val = cookie(res);
 
-            a.getYear().should.equal(b.getYear());
-            a.getMonth().should.equal(b.getMonth());
-            a.getDate().should.equal(b.getDate());
-            a.getSeconds().should.not.equal(b.getSeconds());
-            var delta = a.valueOf() - b.valueOf();
-            (delta > 1000 && delta < 2000).should.be.ok;
-            res.text.should.equal('1');
+            assert.ok(delta > 1000 && delta < 2000)
             done();
           });
         });
@@ -1661,18 +1712,14 @@ describe('session()', function(){
           request(app)
           .get('/')
           .set('Cookie', val)
-          .end(function(err, res){
+          .expect(200, '2', function (err, res) {
             var a = new Date(expires(res))
-              , b = new Date;
+            var b = new Date
+            var delta = a.valueOf() - b.valueOf()
 
             val = cookie(res);
 
-            a.getYear().should.equal(b.getYear());
-            a.getMonth().should.equal(b.getMonth());
-            a.getSeconds().should.not.equal(b.getSeconds());
-            var delta = a.valueOf() - b.valueOf();
-            (delta > 4000 && delta < 5000).should.be.ok;
-            res.text.should.equal('2');
+            assert.ok(delta > 4000 && delta < 5000)
             done();
           });
         });
@@ -1681,15 +1728,14 @@ describe('session()', function(){
           request(app)
           .get('/')
           .set('Cookie', val)
-          .end(function(err, res){
+          .expect(200, '3', function (err, res) {
             var a = new Date(expires(res))
-              , b = new Date;
+            var b = new Date
+            var delta = a.valueOf() - b.valueOf()
 
             val = cookie(res);
 
-            var delta = a.valueOf() - b.valueOf();
-            (delta > 2999999000 && delta < 3000000000).should.be.ok;
-            res.text.should.equal('3');
+            assert.ok(delta > 2999999000 && delta < 3000000000)
             done();
           });
         });
@@ -1708,7 +1754,8 @@ describe('session()', function(){
             request(app)
             .get('/')
             .end(function(err, res){
-              expires(res).should.equal('Thu, 01 Jan 1970 00:00:00 GMT');
+              if (err) return done(err)
+              assert.equal(expires(res), 'Thu, 01 Jan 1970 00:00:00 GMT')
               done();
             });
           })
@@ -1728,7 +1775,7 @@ describe('session()', function(){
             .expect(200, function(err, res){
               if (err) return done(err);
               var val = cookie(res);
-              should(val).not.containEql('Expires=');
+              assert.equal(val.indexOf('Expires'), -1, 'should be not have cookie Expires')
               done();
             });
           })
@@ -1873,10 +1920,16 @@ function cookie(res) {
 }
 
 function createServer(opts, fn) {
+  return http.createServer(createRequestListener(opts, fn))
+}
+
+function createRequestListener(opts, fn) {
   var _session = createSession(opts)
   var respond = fn || end
 
-  var server = http.createServer(function (req, res) {
+  return function onRequest(req, res) {
+    var server = this
+
     _session(req, res, function (err) {
       if (err && !res._header) {
         res.statusCode = err.status || 500
@@ -1891,9 +1944,7 @@ function createServer(opts, fn) {
 
       respond(req, res)
     })
-  })
-
-  return server
+  }
 }
 
 function createSession(opts) {
@@ -1917,6 +1968,29 @@ function end(req, res) {
 function expires(res) {
   var match = /Expires=([^;]+)/.exec(cookie(res));
   return match ? match[1] : undefined;
+}
+
+function shouldNotHaveHeader(header) {
+  return function (res) {
+    assert.ok(!(header.toLowerCase() in res.headers), 'should not have ' + header + ' header')
+  }
+}
+
+function shouldSetCookie(name) {
+  return function (res) {
+    var header = cookie(res)
+    assert.ok(header, 'should have a cookie header')
+    assert.equal(header.split('=')[0], name, 'should set cookie ' + name)
+  }
+}
+
+function shouldSetCookieToValue(name, val) {
+  return function (res) {
+    var header = cookie(res);
+    assert.ok(header, 'should have a cookie header')
+    assert.equal(header.split('=')[0], name, 'should set cookie ' + name)
+    assert.equal(header.split('=')[1].split(';')[0], val, 'should set cookie ' + name + ' to ' + val)
+  }
 }
 
 function sid(res) {